Move countdown in components folder
