+++ front/app/containers/InitiativesShow/ReactionControl/Status/components/CountDown.tsx
@@ -3,7 +3,7 @@ import moment from 'moment';
 import styled from 'styled-components';
 import { colors, fontSizes } from '@citizenlab/cl2-component-library';
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from './messages';
+import messages from '../../messages';
 import { convertSecondsToDDHHMM } from 'utils/dateUtils';
 
 const Container = styled.div`
